refactor(state): consolidate state helpers
