--- apps/campfire/src/state/setImmer.ts
@@ -1,21 +1,0 @@
-import { produce } from 'immer'
-import type { Draft } from 'immer'
-import type { StoreApi } from 'zustand'
-
-/**
- * Zustand's `set` function signature.
- *
- * @template T - The state shape managed by the store.
- */
-type SetFn<T extends object> = StoreApi<T>['setState']
-
-/**
- * Wraps Zustand's set function with Immer's produce for immutable updates.
- *
- * @param set - The original Zustand set function.
- * @returns A function accepting an Immer recipe to update state.
- */
-export const setImmer =
-  <T extends object>(set: SetFn<T>) =>
-  (fn: (draft: Draft<T>) => void) =>
-    set(produce<T>(fn))
